+++ src/views/ZaakDetailView.vue
@@ -14,13 +14,7 @@
       <utrecht-heading :level="1"
         >Zaak {{ zaak.data.identificatie }}</utrecht-heading
       >
-      <a
-        v-if="deeplink"
-        :href="deeplink"
-        target="_blank"
-        rel="noopener noreferrer"
-        >Open in zaaksysteem</a
-      >
+      <zaak-deeplink :zaak="zaak.data" />
     </header>
 
     <tab-list v-model="activeTab" v-if="zaaksysteemId">
@@ -67,9 +61,9 @@ import {
   useContactmomentenByObjectUrl,
 } from "@/features/contact/contactmoment";
 import ZaakPreview from "@/features/zaaksysteem/components/ZaakPreview.vue";
+import ZaakDeeplink from "@/features/zaaksysteem/components/ZaakDeeplink.vue";
 import { TabList, TabListItem, TabListDataItem } from "@/components/tabs";
 import BackLink from "@/components/BackLink.vue";
-import { useZaaksysteemDeeplink } from "@/features/zaaksysteem/deeplink";
 
 const props = defineProps<{ zaakId: string; zaaksysteemId: string }>();
 const contactmomentStore = useContactmomentStore();
